refactor: extract SettingsStore+Models extensions (#636 slice D)
Extract the model-choice cluster from SettingsStore into a new
`SettingsStore+Models.swift` extension file. Continues #636 after
slice A (PR #853), slice B (PR #855), slice C (PR #859).

Moved to the extension:
- 4 accessors: `preferredModelValue`, `issueExtractionModelValue`,
  `transcriptionModelChoices`, `issueExtractionModelChoices`
- 4 static helpers: `transcriptionModelChoices(for:)`,
  `issueExtractionModelChoices(for:)`,
  `normalizedTranscriptionModel(_:for:)`,
  `normalizedIssueExtractionModel(_:for:)`
- 4 constants: `openAITranscriptionModelChoices`,
  `parakeetTranscriptionModelChoices`,
  `openAIIssueExtractionModelChoices`, `parakeetTranscriptionModel`

Visibility changes (minor):
- The 4 static helpers widen from `private static func` → `static
  func` so the property-observer setter path in the main class
  body (aiProvider observer at :92-94 calling
  `normalizeXForCurrentProvider(persist:)`, which calls
  `Self.normalizedTranscriptionModel(...)`) can still reach them
  from a different file.
- 3 constants that are used ONLY by the moved helpers stay
  `private static`. `parakeetTranscriptionModel` widens to
  `static` because it's also referenced by
  `transcriptionModelPlaceholder` at line 546 in the main class.
- 2 constants stay in the main class body but widen from
  `private static` → `static`: `openAITranscriptionModel` and
  `openAIIssueExtractionModel` — both referenced from the moved
  static helpers and from other main-class accessors, so they
  must be reachable across files. `defaultLanguageHint` stays
  `private static` in the main class.

Zero external test callers of any widened symbol (`grep Tests/`
confirmed).

Line-count delta:
- SettingsStore.swift: 1807 → 1698 (-109)
- New SettingsStore+Models.swift: 113 lines
- xcodeproj: +4 (xcodegen picks up the new file)

Tests: `xcodebuild build` + `SettingsStoreTests` pass; the
existing tests at Tests/BugNarratorTests/SettingsStoreTests.swift
lines 484 and 510 (`store.transcriptionModelChoices.map(\.id)`,
`store.issueExtractionModelChoices`) verify the moved accessors
work unchanged.

Closes #862.
